add math support to articles
The math uses KaTeX to statically generate the math (via a {% latex %} liquid tag). The CSS stylesheets are enabled by using a has_math variable in the Jekyll --- thingy (so they aren't loaded unnecessarily). This has a potential to be automated -- the CSS will only be added once a {% latex %} tag is used.