add support for dark theme
I found out that there is a super easy way to implement theme switching for a website. This pretty much solved all my problems with not being able to do this without JS! I'll write an article about it later, since the information about how to actually turn this on in FireFox was quite murky.