Reduce xopp svg sizes
I had to rewrite a bunch of stuff, because export from Xournal++ directly to SVG is broken atm. This means that we're converting to PDF, then to SVG using Inkscape, then do some regex magic to remove the background and save it, then call Inkscape to crop to the drawing and finaly use svgcleaner to reduce the size. I want to die.